+++ core/model/src/main/java/com/mmorikawa/core/model/Rating.kt
@@ -4,7 +4,7 @@ import kotlinx.datetime.Instant
 
 data class Rating(
     val book: SimpleBook,
-    val score: Int,
+    val score: Float,
     val ratingText: String,
     val dateCreated: Instant,
     val dateUpdated: Instant,
